requirements.txt
python-dotenv==0.6.4
spotipy==2.4.4
fire==0.1.1
requests==2.14.2